refactor: use baseUrl at tsconfig
