+++ src/components/pages/Auth.tsx
@@ -1,9 +1,10 @@
 import React, { useEffect } from 'react';
 import { Result, Typography } from 'antd';
 import { LoadingOutlined } from '@ant-design/icons';
-import { IAuth } from './IAuth';
-import { getToken } from '../../util/ft_api';
-import history from '../../history';
+
+import history from 'util/customHistory';
+import { IAuth } from 'components/pages/IAuth';
+import { getToken } from 'util/ft_api';
 
 const { Text } = Typography;
