+++ src/index.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+// import { Provider } from 'react-redux';
 import './index.css';
 import { createGlobalStyle } from 'styled-components';
-import App from './components/App';
+import App from 'components/App';
+// import store from './util/redux/store';
 
 const GlobalStyle = createGlobalStyle`
   html, body {
